pyproject.toml
[tool.poetry]
name = "duckdb-llm-cli"
version = "0.1.0"
description = "CLI utility for natural language queries to DuckDB using LLMs"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
duckdb = "^0.9.2"
openai = "^1.3.5"
pandas = "^2.1.3"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
duckllm = "duckdb_llm_cli.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
